feat(crud): harden Players and Comments CRUD with cascade delete, validation, concurrency, rate limit and soft-delete
Tightens every gap surfaced in the post-CRUD review and ships the API behind WebApplicationFactory integration tests.

Tier 1 — close loose ends:
- DeletePlayerHandler cascades to Comments.API via CommentsClient.DeleteByPlayerIdAsync (Polly Circuit Breaker covers Comments outages — orphaned comments fall back to logged warning).
- SearchPlayersHandler escapes %, _ and \ before passing patterns to EF.Functions.ILike so user input cannot break out of the LIKE.
- CreatePlayerHandler returns 400 on duplicate (Season, TeamName, LeagueName) inside the Statistics payload before EF hits the UNIQUE constraint.
- GET /api/players/{id} now emits _links.comments as relative /api/comments/player/{id} (resolved through the gateway).

Tier 2 — validation:
- CreatePlayerValidator/UpdatePlayerValidator: Name MinimumLength 2.
- BirthDate added to Create/UpdatePlayerDto, validated to be 10-60 years old.
- Comment.Rating migrated from int to decimal with step 0.5: domain rejects non-half values, FluentValidation enforces InclusiveBetween + step, EF column is now numeric(2,1). DB check constraint dropped because Microsoft.Data.Sqlite serialises decimals as TEXT and would break SQLite tests; defense-in-depth lives in domain + validator.

Tier 3 — production hardening:
- Optimistic concurrency: Player.Version (portable int instead of xmin/rowversion to keep SQLite tests working). PlayersDbContext bumps Version on Modified entries; PlayerRepository wraps DbUpdateConcurrencyException and rethrows ConcurrencyConflictException (Shared). UpdatePlayerHandler reads optional If-Match header → 412 Precondition Failed on mismatch, 409 Conflict on EF race. Controllers expose Version in PlayerDetailDto and as ETag header.
- DELETE is idempotent: Player and Comment handlers return 204 unconditionally.
- POST /api/comments/player/{id} rate-limited to 5/min per X-User-Id (or IP) via Microsoft.AspNetCore.RateLimiting fixed-window policy.
- Soft-delete on Player and Comment: DeletedAt nullable column, HasQueryFilter hides deleted rows, DeleteAsync calls MarkDeleted() + SaveChanges. DeleteByPlayerIdAsync rewritten as ExecuteUpdateAsync to set DeletedAt. Migrations added.
- Integration tests: FootballManagerApp.Players.API.Tests and FootballManagerApp.Comments.API.Tests bring up the real WebApplicationFactory<Program> with SQLite swapped in for Aspire's Postgres registration. Cover happy path POST→GET→PUT→DELETE plus 401/403/409/400/304-style cases. WAF teardown clears the EF service provider before registering SQLite.

Build: 0 warnings, 0 errors. Tests: 81/81 across 7 projects (Gateway 1, Comments.Application 21, Players.Application 27, Comments.Infrastructure 9, Players.Infrastructure 15, Players.API 4, Comments.API 4).

+++ src/FootballManagerApp/FootballManagerApp.Comments.Application/Comments/DTOs/CommentDtos.cs
@@ -5,13 +5,13 @@ public record CommentDto(
     Guid PlayerId,
     string Author,
     string Text,
-    int Rating,
+    decimal Rating,
     DateTime CreatedAt);
 
 public record CreateCommentDto(
     string Author,
     string Text,
-    int Rating,
+    decimal Rating,
     decimal? ClientLat,
     decimal? ClientLng,
     string? ClientCity,
